CoreFAPI/BackApp/search/210325/TRUE/210325_3R/search_service/requirements.txt
fastapi
uvicorn
elasticsearch==7.17.9
httpx
pydantic
pydantic-settings
python-dotenv
redis
aiohttp